pyproject.toml
[project]
name = "uv-pdf2csv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "pandas>=2.2.3",
    "python-multipart>=0.0.20",
    "tabula-py>=2.10.0",
    "uvicorn>=0.34.2",
]
